requirements.txt
streamlit
pydub
openai
whisperlib
google-gemini-api
python-dotenv
ibm-watson